Add files via upload
This upload contains some Colab files about the public data analysis. "Copie of Twitter API" is the Colab file responsible for collecting data from Twitter. "Copie de Google Trend & Key Dates" is a graph analysis of the evolution of public interactivity with the topic of genetically modified crops along with key dates in the history of genetically modified crops. "Pearson GM_PO" file is a Pearson correlation analysis on a national level, between general United-States GM Production, and public interactivity with the topic.

Note that the file "Cleaning Twitter API" requires a pre-cleaned string of the Twitter API results to work. Inès provided a better Colab file which handles a CSV file format to clean the Twitter API Data and to process to its analysis.